+++ types/server.go
@@ -44,6 +44,7 @@ type Server struct {
 	Short                  string             `db:"short" json:"short" description:"The server's short description"`
 	Long                   string             `db:"-" json:"long" description:"The server's long description in raw format (HTML/markdown etc. based on the servers settings). May not be included in responses (e.g. long is not set in include)" skip:"long" ci:"internal"` // Must be parsed internally
 	Type                   string             `db:"type" json:"type" description:"The server's type (e.g. pending/approved/certified/denied etc.)"`
+	Note                   pgtype.Text        `db:"approval_note" json:"approval_note" description:"The note for the server's approval"`
 	State                  string             `db:"state" json:"state" description:"The server's state (public, private, unlisted, defunct)"`
 	Tags                   []string           `db:"tags" json:"tags" description:"The server's tags"`
 	VanityRef              pgtype.UUID        `db:"vanity_ref" json:"vanity_ref"`